+++ src/Domain/SubmitDeltaXmlFileResponse.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DjThossi\ErgosoftSdk\Domain;
+
+readonly class SubmitDeltaXmlFileResponse
+{
+    public function __construct(
+        public StatusCode $statusCode,
+        public JobGuid $jobGuid,
+        public SubmitDeltaXmlFileResponseBody $responseBody,
+    ) {
+    }
+}